tests: update test_fair_share for integration tests
This patch adds a simple test where demo_fair_share is run
for the fifo and grouped_read I/O schedulers. It sends requests
continuously for 5 seconds and prints the success/failure statistics
of the requests sent by the demo_fair_share.

Also add a timeout to tail in waive_daemon so that the daemon is not
stuck if some requests are pending on the client.

Change-Id: I671b4f67f7c32004b3713428406a0e5fcb3a656f
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11582
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>