doc: specify what the replica count means
This patch updates the documentation for the raid1 layout API and
configuration file to specify what exactly Phobos implies when we talk
about the "replica count".

This value is the number of data replicas, so a replica count of 1
means that there is only one copy of the data (the original), and 0
additional copies of it. Therefore, a replica count of 2 means the
original copy of the data, plus 1 additional copy.

Change-Id: Id4ee1a847535161d92aaf67ee80155c1f5c6de7f
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11655
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>