db: add 2.0 schema to prepare extent migration
This database schema update is the first step of the 'extent migration'
feature. It mainly consists in three modifications:
- changing the granularity of the extent table, from one object to one
extent per row;
- bursting the json information for an extent in multiple columns;
- adding a layout table to bind an object with its extents.

A test is added to check an object and its extents information are
correctly scattered through the three tables when migrating from 1.95 to
2.0.

For now, the default schema is still the 1.95 as there will be huge
modifications in the source and tests code to consider the new schema.
The update of the database.md design document will come once the 2.0
schema is set as the default one.

Change-Id: Ia4597e5678ea9f32b698b384ed44f92f80c9ce25
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11669
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>