cli: fix `--degroup` option for `phobos extent list`
This patch fixes an issue where the output of the command line was not
correct when using the `--degroup` option. In this case, if an object is
composed of multiple extents, they are all described with the first
extent information.

This patch also adds a test to check the outputs of simple `phobos
extent list --degroup` commands are correct.

Change-Id: I5528aad40ead54220eeb92dbfb68cf7b92236004
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11688
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Alexis GULLO <alexis.gullo.ocre@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>