doc: fix fair share documentation
The exemple documentation is installed with the RPMs. This configuration
limits the maximum number of formats for LTO5 tapes to 0 which means
that by default, no drive will be able to format an LTO5 tape. Since LTO5
and LTO6 tapes are used for testing, this patch uses LTO3 and LTO4
instead. It leaves sensible defaults for LTO5 and LTO6.

Also move test_fair_share.test to comply with style guidelines.

Change-Id: I28dbf8ee69256724d12367d62e05a151cb7794e8
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11702
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>