logs: add message logging to library open at device load
This patch adds logging in case a library open fails at device load.
Other cases where a library open is needed will be managed in other
patches.

Also fix a bug where if the lib open in `dev_load` would fail and
`release_medium_on_dev_only_failure` was true, we would try to release
the medium held by the device, which is NULL at this point, resulting in
a segfault by the LRS.

Change-Id: I460c28d5abc1bc0d64b7fcdf5632889a3f8f5cb7
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11758
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>