tape_import: added cli command for tape import
This patch adds the cli command for the tape import feature.

This patch also adds the 'importing' status for media, used during
the tape import function, as a way to add the tape to the dss with
a different status from 'blank'

This patch modifies the previous `add` CLI function (used in tapes)
to add a new API `tape add` version of it, that basically adds
a new entry to the DSS.

This patch also adds the tape that is to be imported to the DSS
with the "importing" fs_status and with the PUT, DELETE flags to
false.

Change-Id: I0018c8dddf17d5220dee848ec01b9e4854b378da
Signed-off-by: Alexis GULLO <alexis.gullo.ocre@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11782
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>