logs: add logging to move target selection
This patch adds logging to the target address selection call done when
trying to move a medium in `dev_load`. It thus modifies the
`lib_scsi_move` function to instead create two JSON sub-objects
regarding target selection and medium movement.

With this patch, all SCSI interactions done for a device load are
logged.

Change-Id: I811db2f862a658118e789c696b5a49f0800bc2e5
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11760
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>