tests: prepare tests for scsi logging
This patch aims to prepare the SCSI logging tests that will be done by
mocking the "ioctl" calls in the "scsi_execute" function. This is done
through a callback specified in the global Phobos context, and mocking
functions that set by the test.

Currently this test does not verify much as it needs a way to retrieve
the logs, which will be implemented in a later patch.

Change-Id: I90d40a7fc518c251a7a9f9d67f333e058e455db3
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11793
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>