tests: update intern test for SCSI logging
This patch updates the intern test that check the SCSI logging in done
correctly after a call to "dev_load". It checks that if any interaction
done with the library fails, the logs are correctly filled and emitted.
Also check that if the "dev_load" operation succeeds, the log is
correctly filled aswell.

Change-Id: Id6384ce1d50be30fc8d94b788d1c0db63a704efd
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11823
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>