tlc: library file descriptor management
This patch introduces the tlc lib_device option in phobos
configuration to provide to the TLC the path of the tape library
device.

The tape library file descriptor field is added into the tlc
structure.

The TLC opens the tape library file descriptor at init and closes
it before stopping.

Change-Id: I7fe96e22d90ec4a5e6a576d06036e11227941f49
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11879
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>