io-sched: fix invalid count of devices per scheduler
When there are fewer devices available than the number of schedulers in
need of a device, the previous implementation would count the total
number of devices per scheduler to determine how many devices need to be
distributed to a given scheduler.

This behavior is not correct, we need to count device per tape
model since they are distributed independently.

Change-Id: I7226294316718ad0ec32d888544d43bfb3ad8a73
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11950
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>