lrs: fix unlock concurrence between sched and device threads
At write, the scheduler was taking a lock early on the selected
medium and then unlocking if no dev was later found. This later
unlock was conflicting with the device thread's lock
in heavy concurrent situation.

This patch moves the locking at write medium selection only
before passing the request to the device thread. This
removes the need for sched unlock on medium selection at
write.

Change-Id: Ib076ea458d8a7da08b8885bdfb68c2e07e607dbc
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11956
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>