lrs: fix compatible_drive_exists invalid list usage
compatible_drive_exists needs a list of devices used during an
allocation. This list is stored in resp_container::devices. The caller
of this function was dereferencing the list before passing it to
compatible_drive_exists and compatible_drive_exists would assume that
the pointer given as a parameter selected_devs was the start of a list
of struct lrs_dev. This resulted in an invalid use of the list.

Change-Id: If55ac20e6e09bfab10efa33cdfbb85380ab79cc6
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11961
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Tested-by: Patrice LUCAS <patrice.lucas@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>