obj_reconstruction: add the function of reconstruction for an object
This patch adds the feature of reconstruction of an object after an
import. After an import, the extents of the `incomplete` and
`readable` object can be browsed to update their status depending
on the number of copies of the object found.
Also, this patch adds the offset attribute inside of the `extent`
json attribute of the `extent` table, used in the reconstruction
mecanism to detect the last split of an object.

Change-Id: I57610344deae40791c94ed75b338cf0124ace762
Signed-off-by: Alexis GULLO <alexis.gullo.ocre@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12022
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>