lrs: release device thread ongoing_io status on comm error
When an error occurs when sending an answer, we never try to
send again this answer message and the answer message is
definitely lost.

If this answer was a read or write allocation, we need to clear
the ongoing_io status of the corresponding device threads because
the corresponding client will never get this allocation answer
and will never release it in the future.

Change-Id: I9a79b8e9c585fd1734114df7a581c346c2fcfc4b
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12081
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>