lrs: clean lock file if daemon start fails
If the DAEMON_PID_PATH environment variable is not defined,
the parent process fails and closes the pipe it shares with the
child daemon. Closing this pipe fails the child. In this case,
the lrs child daemon must clear the lock file before exiting.

This patch also adds the lock_file path into the lrs structure.

This patch also manages the SIGPIPE signal when the child daemon
tries to write its return code to the father daemon.

Change-Id: I878c8d3d20576cb154f3b256ac6619211e349412
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12083
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>