lrs: fix medium lock concurrency issue between device and sched
We can't rely on duplicated medium lock info. Before this patch,
some media info structures were early loaded with existing lock
info in schedulers. This lock was never refreshed by the
scheduler whereas it could be cleaned by concurrent threads.
For example, medium are unlocked at unload by device threads.
This could happen on behalf of the scheduler thread.

With a wrong remanent existing lock, the scheduler was not taking
this lock before pushing the corresponding request to the device
threads. It was leading to load medium without locking them
again. At unload, the device thread was failing because it tried
to unlock a medium that was not locked.

This patch also adds the corresponding test.

Change-Id: Ic5eef91f8182cdd193f5c1a81f30af3e17364ef2
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12127
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>