logs: add all filters to logs clear
This patch adds all filters created for the "logs dump" command
to the "logs clear" command. Also move "dss_logs_delete" since it now
uses a static function in dss.c (may be changed later, but not necessary
for this patch).

A simple test is added here, but more will be in the next patch.

Change-Id: Iad401fffbbe6056decee4de214447148823f82b8
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12130
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>