tlc: introduce future tlc_library code
This patch is a "fake" one only to introduce the future code
of the tlc library. This code is copy/past one from the existing
ldm_library code. So to emphasize modification of the code and ease
the review process, we pre-add a commented copy of the file that
will be modified all along the TLC task.

Change-Id: I6464d237dfe0420ad03a1348f9f1352992d2d9e6
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12185
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>