tests: fix an issue between test_store and test-driver
This patch fixes a problematic interaction between a test in
test_store.test, the test-driver and the log file for test_store.test.
The test in question is the following:
"$test_bin tag-put file mytag |& tee /dev/stderr | \
     grep -e '/tmp/pho_testdir[245]'"

This command would run "$test_bin" with the given parameters, and pipe
the stderr of that command to a "tee". This "tee" would redirect the
input (so "$test_bin"'s stderr) to the next pipe (the grep which
verifies the command did what is was expected to) and the file
"/dev/stderr", which is a symlink to the standard error.

On the other hand, the test-driver behaves like this when running a test
function:
"$test_func >> $log_file 2>&1"

Here, three things happen: first the test function is executed, then
it's standard output is redirected to a log file, and finally, the
standard error is redirected to the standard output. This means both
stderr and stdout are redirected to the log file in the end.

Therefore, when the first test is run with the test-driver, the "tee"
will redirect the standard error of "$test_bin" to the standard error
(again), but since the test-driver redirects the standard error to
log_file, ultimately, the "tee" redirects the standard error to the log
file.

However, that is problematic because the log file is filled
progressively with the test's content, but the "tee" command, if given
an already existing file, will overwrite it's content. So since the
"tee" redirects the standard error to the log file, which already
exists, it will overwrite its content, reseting the log file's size to
0, thus removing the log for that test (here, test_store.test).

So this patch fixes this issue by simply removing the tee command, since
the actual verification done by that command (the grep), only need to
have it's standard input filled, which "|&" already does, thus the "tee"
is unnecessary.

Change-Id: I60321d1636f686dce505be8c767bb83b0f5e6bcc
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12304
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>