common: add xmalloc, xstrdup, xcalloc...
These functions will abort in case of an allocation failure.
Most of the time, the only error a function returns is an allocation
failure. Since they are not likely to fail on Linux, it has been decided
to avoid the checks when possible.

Moreover, the current code tries to check for every allocation failure
but never does anything appart from stopping whatever action was
occurring. So aborting has the same behavior in practice.

Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Change-Id: I67dbb9823b6f4aa129dab086b353480791ae206b
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12383
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>