ltfs: add logs to the dev_mount function
This patch adds logs to the dev_mount function in the LRS. It is similar
to the work done for the SCSI logs, but here the File System adapter
LTFS is modified, specifically the "ltfs_mount" function (which is
called by dev_mount"). The other calls in dev_mount will be updated
aswell.

Alongside these modifications, this patch creates a new schema for the
2.0 with the relevant operation type added ("device mount") and a first
test of the "ltfs_mount" function. It is done similarly to the SCSI logs
test, in that we add mocking function in the Phobos context that will be
called by "ltfs_mount". They normally are the function of the old
version, but are modified in the test to fail when called.

Change-Id: I94e5aef6d9b3457e13d7a91ab57047af3dd99449
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12414
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>