benchmark: add pho-bench script to run benchmarks
$ pho-bench <backend> <benchmark file> [--check] [-d|--debug]

This script will accept a file of the form:

$ cat bench.txt
PUT filename1 P00000L5
GET filename2 P00000L5
PUT filename3 P00001L5

It will run the given actions using the backend (phobos, ltfs, tar or dd).

The script also accepts a --check argument to check that the files that
are written to the backend are the same when read back.

Change-Id: I9512cf2d34743fc2babeac28226e2f8d0908ec84
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12445
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>