lrs: no medium search if no available device
If no device is available, we don't search for a medium.
When all devices are already booked, it's better to stop
instead of finding a "new" available medium because we prefer
to write on an already mounted medium when it will become
available.

This patch also fix the fact that we have compatible devices
even if we have not already selected a medium.

Change-Id: Ieb9641ec8a211cfc2db4fc28cb3f72cd0c8e1e46
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12442
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>