dss: add emit_log_after_action
This function will handle all the logic of emitting a log only when
necessary and formatting it properly before. It will also free the
internal JSON elements.

Also, pass pho_ids by reference instead of by value to init_pho_log as
pho_id is a big struct which is expensive to copy:

$ pahole -C pho_id .libs/libpho_dss.a
struct pho_id {
        enum rsc_family            family;               /*     0     4 */
        char                       name[256];            /*     4   256 */

        /* size: 260, cachelines: 5, members: 2 */
        /* last cacheline: 4 bytes */
};

Change-Id: I0459d7731ed973e1b607f35d825847924e785f9e
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12496
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>