io_sched: use the retry mechanism for write allocations
The write allocation code was not using the retry mechanisms. This means
that any time a drive fails to load a medium for a write, the code will go
through the regular allocation mechanisms which would fail because the
schedulers check that the request we are trying to find a medium
and a device for is the request currently being scheduled (i.e. the
one returned by peek_request).

Since the request was already removed prior to being pushed to the
device, and we are rescheduling the request, it was not returned by
peek_request so the scheduler will return -EINVAL on allocation.

The consequence of that is that if a device fails to load a medium for a
write allocation, the scheduler won't be able to retry with a new medium
as it should.

This patch also introduces some debug logs to follow the request state.

Change-Id: I438661c70e2e5d84e2c1daca0aa88340db222d22
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12493
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>