lrs: remove free_medium parameter to fail_release_free_medium
This parameter is removed to simplify the health management of media.
A medium will need to be set to failed and unlocked when its health
reaches 0 but we won't free it in all the cases. So the logic of freeing
the medium is outsourced to the caller of fail_release_free_medium which
is now called fail_release_medium.

Change-Id: I1d6d0fc69440d020c34bbe1111e2948534558dcb
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12898
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>