lrs_device: fix format cleaning
Once the medium is loaded after calling dev_load, it is
registered into the device structure and must not be freed
by anything other than this structure. For example,
if we face an error at dev_format, after the dev_load, we
must not free the medium to let the device thread clean its
loaded medium during its final cleaning process.

Change-Id: I25b6a4ea65bfaa382e14cffa2d808f0f324b7fa1
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12650
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>