test: fix concurrent_put
This test requires 2 drives/dirs. If not enough are available, the test
will run with only one. The issue is that the test was not supporting
the case with only one drive/dir.

We also update the ensure_nb_drives function to make sure that it
properly counts the number of unlocked devices. It was previously
increasing the number of devices after the unlock be the code was
written in a way that did not ensure that the device was locked prior
to the unlock.

Change-Id: Iac020129367f76db6d2de8e500b72e784769de31
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12671
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>