lrs: do not set sub request medium to NULL if already loaded
When the scheduler searches for a medium/device pair to allocate a
request, it sets to NULL any medium that is already loaded. This
behavior has several advantages:
- the device thread can simply look at the medium pointer and if NULL it
  knows that we want to use the currently loaded medium;
- we don't need to worry about freeing the medium when already loaded.

The downside of this approach is that the sub request is not sufficient
to have access to the medium we are trying to use. We also need to have
a reference to the lrs_dev structure. This forces an unnecessary
dependency with lrs_dev in parts of the code that just want to access
the medium's information.

So this patch removes this feature and copies the pointer to
lrs_dev::ld_dss_media_info when appropriate so that the request always
contains the full list of media to allocate.

This patch changes the behavior of the device thread's handling of
formats. In the case of the format --force option, if the medium is
already in the state "mounted", we previously would unmount LTFS and
unload the tape then reload it then do the format. This behavior was
suboptimal. Since LTFS cannot format a mounted filesystem, we need to
unmount the device first before the format.

Change-Id: I9599f41ef7af5d8a467288fb93922fee0563568d
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12899
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>