lrs: signal the device thread when scheduling a format
The device thread will sleep until the next sync is scheduled to avoid
unnecessary loops that will do nothing. This is also done when the
request is a format. But the function `sched_handle_format` did not
signal the device thread. This means that the device thread will sleep
for at most `sync_time_ms` before starting the format. We add the signal
to avoid unnecessary waiting.

Change-Id: I7a4ec662d7d9d769694dbbda224ae72a6e3bfae6
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12673
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>