tests: remove test_locate_locked_splits for tape family
I don't understand how this test could work in the first place.
The media where not properly updated by resize_medium since the variable
$tapes was of the form P0000[0-7]L5 which was not interpreted correctly
by the function. Also, since the LRS updates the size available after a
write is finished, at the end of the write of the first split, the
tapes' sizes are set back to their correct values. So the second split
is still written on the same 4 media used for the first split.
Finally, the test was updating the devices' hostname to test the locate
feature. But it was using the medium name as the device's path in the
SQL query. This only works for dir and rados_pool families.

Since this test does not make sense, the test is removed for tapes.

Change-Id: Ic5f7ee22854e80ae6e5276cef871d1a082f58226
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12687
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>