common: add a generic cache mechanism pho_cache
This patch adds a new data type pho_cache that will cache values of
arbitrary types. Values are indexed by a key that is needed to get a
reference. Values in the cache can be managed using the following
functions:

- pho_cache_acquire: get a value from the cache given a key.
  If the value is not in the cache, it is built using
  pho_cache_operations::pco_build. Otherwise, the value in the cache is
  returned and its reference count is increased.
- pho_cache_release: release a value to the cache.
  The reference count of the value is decreased and if it reaches 0,
  the value is freed using pho_cache_operations::pco_destroy.
- pho_cache_insert: add a value that was built externally to the cache
  (i.e. without a call to pho_cache_operations::pco_build). This will
  be helpful to build the media cache in the LRS. The LRS will fetch
  media from the DSS to select the best one for a write allocation.
  Once the best one is selected we can simply insert it in the cache
  without an extra DSS request.
- pho_cache_update: explicitly build a new value and replace it in the
  cache to get the most up to date version and return it. This
  function ensures that the value we receive is up to date by creating
  a new value with pho_cache_operations::pco_build.

The cache allows the user to insert values built externally. This is
because the LRS can fetch values from the DSS and select the most
optimal one for an I/O. In this case, we don't want to make a second DSS
request simply to retrieve the medium info that was already fetched.

We also can update values that we already have a reference to because in
some parts of the LRS' code, we need to detect if a medium was locked
for instance.

Both of these constraints imply that some thread may have an outdated
reference to the cached value. But since they still hold a reference to
it, it needs to be kept in the cache in order to be released when the
reference count becomes 0. This property makes it easier to manage
cached values from different threads. As long as a thread has a
reference to a value, it knows that the structure won't be freed and
other threads can update the value in the cache without risking to free
the old value that another thread is currently holding. This is
especially interesting for the media cache since most of the time the
medium info of the device thread (lrs_dev::ld_dss_media_info) is
accessed, it is to read invariant data such as its ID or FS type.  Since
they don't change, reading it from an older version of the media_info is
not an issue. This avoids taking extra locks each time we want to print
the name of the medium for example.

The cache has been designed so that values are retrieved using a key and
released using the pointer to the value that was returned. This makes it
easier for users of the cache to access and release references. But this
forces the implementation to quickly have access to an element of the
cache by using either the key (pho_cache_acquire) or the value
(pho_cache_release). To achieve this, the key value pair is stored in
one structure (struct key_value). This structure will wrap the value and
prepend a pointer to the key to it.

Therefore, container_of(value, struct key_value, value) will give us
back the struct key_value from which we can find the key. This imposes a
restriction on how the values are built. They must be embedded in the
key_value structure. To this end, a callback
pho_cache_operations::pco_value2kv is necessary in some cases to convert
a value to a key_value.

Change-Id: I77b879e2b6a077116fc8af51fab3a2cf4f52bdab
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12743
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>