lrs: use media cache
This patch introduces the media cache in the LRS. Since the cache is
shared between all the threads (communication, device and scheduler),
extra care must be taken when accessing a media_info.

Several atomic operations are added to manage the device thread's
medium. The device thread has its own reference on the loaded medium.
Since it can drop this reference when unloading the tape, another thread
that wants to access this structure must have their own reference. This
prevents the risk of having the device thread free the media_info when
other threads are still using it.

The other issue is that the device thread will set the medium to NULL
when unloading it. This means that the pointer can become NULL at any
point in other threads. To prevent this issue, the reference to the
device thread's medium must be taken with the device lock held.

The free_medium argument of dev_load is removed. Now that we use the
cache, there are references on the medium of a sub request:
- one in the req_container;
- one in the device thread once the medium is loaded.

The free was done to remove the media_info from the req_container as
they are loaded into their device. This is not necessary since all the
media_info are released when the req_container is free'd. The downside
is that the reference must be released when a sub request is sent back
to the scheduler after a failure. Therefore, the release is also added
where necessary in the I/O schedulers.

Change-Id: I487140eb1e1b00c5a7c2db0c443540c3ecb9fa21
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12753
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>