lrs: add read_media_list structure
For read allocation requests, the client will send a list of media that
the LRS can choose from. Once the scheduler chooses a medium to
allocate, its media_info will be stored in the rwalloc_params' list of
media.

This list and the list of pho_rsc_id_t in the read allocation request
must be ordered in the same way. This means that when we want to
allocate the i-th medium of the list, we need to swap the corresponding
pho_rsc_id_t to the i-th position in the read request's list.

This patch introduces an intermediate structure that will hide this
complexity from the caller. The caller will have 3 ways to indicate
how to swap a given medium:
- `rml_medium_update(list, index, RMAS_OK)`: the medium can
  successfully be sent to the corresponding device thread;
- `rml_medium_update(list, index, RMAS_UNAVAILABLE)`: the medium cannot
  be sent to the corresponding device thread due to a temporary
  unavailability (e.g. loaded in a busy drive). It won't be used for
  this allocation but may be considered again if another medium cannot
  be used or when the allocation is retried later;
- `rml_medium_update(list, index, RMAS_ERROR)`: an error has occured
  that makes it impossible to use this medium now or later (e.g.
  administrative lock, failed status, ...).

This structure might also be useful in the future to implement
heuristics for the selection of the medium to use for a read
allocation. Given the list of media that can be used, which one is the
best? For example, the one with the most requests associated to it.

Change-Id: I9012c51389476098bfb72212433f7056a559cc42
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12740
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>