lrs: change medium parameter of dev_load to simple pointer
dev_load was taking a reference to a pointer to the medium to load. The
reason for this was that the medium was freed in case of error. Since
the caller doesn't need to access or modify the medium to load after a
call to dev_load, no need to set it to NULL.

Change-Id: Ic94dbd3df3a07254e41bffdecf88b9f439c8b059
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12901
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>