lrs: add health management for device and medium
The health of a medium or device is cached in memory. It reflects the
number of successes/failures in the DSS for the given device/medium.
This counter must therefore be aligned with the state of the DSS.

When an operation succeeds, increase the health of the medium and the
device. When it fails, decrease it. The operations supported are:
- load/unload
- mount/umount
- sync
- format

A successful load -> mount will only increase the health by 1 since on
error, the health is decreased by 1 as well. The health increase must be
symmetric to the decrease otherwise, there will be an imbalance between
the health stored in the DSS and the one the LRS has in memory.

Change-Id: I1e36a5ef8bcafa67b468968ec4078b9a844d6022
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12902
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>