admin: add phobos_admin_notify_media_update
This function will notify the LRS when a medium or a list of media is
updated in the DSS from the CLI if the medium is locked by the local
LRS. If the function is executed on the same node as the LRS, a
notification message will be sent and the LRS will update its internal
state accordingly. Otherwise, a warning will be display to the user to
indicate that the medium was not updated in the owners cache. This means
that the value currently in the cache will not be up to date and the
admin will have to restart the LRS daemon if they want it to use the
most up to date value.

A new CLI command can easily be added later to notify the LRS
accordingly. For example:

> phobos dir lock /mnt/dir
<WARN> Medium '/mnt/dir' is locked by 'somehost', it will not be notified..
> ssh somehost phobos dir notify /mnt/dir

Change-Id: Ib4af0427adcd66541ae2740eea692d9e3c623faf
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12748
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>